refact mvp and mvp changer
