+++ core/utxo/src/mvp.rs
@@ -1,72 +1,38 @@
-use sr_primitives::traits::{Zero, CheckedAdd, CheckedSub};
+use super::*;
 // use Encode, Decode
 use parity_codec::{Encode, Decode};
-use std::ops::{Deref, Div, Add, Sub};
-use serde_derive::{Serialize, Deserialize};
+use plasm_merkle::{MerkleTreeTrait};
 
-#[derive(Clone, Encode, Decode, Default, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Clone, Eq, PartialEq, Default)]
 #[cfg_attr(feature = "std", derive(Debug))]
-pub struct MVPValue(u64);
+pub struct MVPInserter<T: Trait, Tree: MerkleTreeTrait<T::Hash>>(PhantomData<(T,Tree)>);
 
-impl MVPValue {
-	pub fn new(a: u64) -> Self {
-		MVPValue(a)
-	}
-}
-
-impl Div<usize> for MVPValue {
-	type Output = MVPValue;
-	fn div(self, rhs: usize) -> Self::Output {
-		MVPValue(*self / (rhs as u64))
-	}
-}
-
-impl Zero for MVPValue {
-	fn zero() -> Self {
-		MVPValue(0)
-	}
-
-	fn is_zero(&self) -> bool {
-		**self == 0
-	}
-}
-
-impl Add for MVPValue {
-	type Output = Self;
-	fn add(self, rhs: MVPValue) -> Self::Output {
-		MVPValue(*self + *rhs)
-	}
-}
-
-impl CheckedAdd for MVPValue {
-	fn checked_add(&self, v: &Self) -> Option<Self> {
-		if let Some(v) = (**self).checked_add(**v) {
-			return Some(MVPValue(v));
+impl<T: Trait, Tree: MerkleTreeTrait<T::Hash>> Inserter<T> for MVPInserter<T, Tree> {
+	fn insert(tx: &T::Transaction) {
+		Self::using_insert(tx);
+		let hash = <T as system::Trait>::Hashing::hash_of(tx);
+		for (i, _) in tx.outputs().iter().enumerate() {
+			Tree::push(<T as system::Trait>::Hashing::hash_of(&(hash, i)));
 		}
-		None
 	}
 }
 
-impl Sub for MVPValue {
-	type Output = Self;
-	fn sub(self, rhs: MVPValue) -> Self::Output {
-		MVPValue(*self - *rhs)
-	}
-}
-
-impl CheckedSub for MVPValue {
-	fn checked_sub(&self, v: &Self) -> Option<Self> {
-		if let Some(v) = (**self).checked_sub(**v) {
-			return Some(MVPValue(v));
-		}
-		None
-	}
-}
-
-impl Deref for MVPValue {
-	type Target = u64;
-
-	fn deref(&self) -> &Self::Target {
-		&self.0
-	}
-}
+//#[derive(Clone, Eq, PartialEq, Default)]
+//#[cfg_attr(feature = "std", derive(Debug))]
+//pub struct MVPRemover<T: Trait>(PhantomData<T>);
+//
+//impl<T: Trait> Remover<T> for MVPRemover<T> {
+//	fn remove(tx: &T::Transaction) {
+//		Self::using_remove(tx);
+//	}
+//}
+//
+//#[derive(Clone, Eq, PartialEq, Default)]
+//#[cfg_attr(feature = "std", derive(Debug))]
+//pub struct MVPFinalizer<T: Trait, Tree: MerkleTreeTrait<T::Hash>> (PhantomData<T>);
+//
+//impl<T: Trait, Tree: MerkleTreeTrait<T::Hash>> Finalizer<T> for MVPFinalizer<T, Tree> {
+//	fn finalize(authorities: &[<T as consensus::Trait>::SessionKey]) {
+//		Self::using_finalize(authorities);
+//	}
+//}
